@@ -21,7 +21,7 @@
 package eu.europa.ec.dgc.gateway.client;
 
 import eu.europa.ec.dgc.gateway.model.JrcRatValuesetResponse;
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
 import org.springframework.validation.annotation.Validated;
