@@ -22,10 +22,10 @@ package eu.europa.ec.dgc.gateway.repository;
 
 import eu.europa.ec.dgc.gateway.entity.RevocationBatchEntity;
 import eu.europa.ec.dgc.gateway.entity.RevocationBatchProjection;
+import jakarta.transaction.Transactional;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
-import javax.transaction.Transactional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -44,14 +44,18 @@ public interface RevocationBatchRepository extends JpaRepository<RevocationBatch
     Optional<RevocationBatchEntity> getByBatchId(String batchId);
 
     @Modifying
-    @Query("UPDATE RevocationBatchEntity r SET r.signedBatch = null, r.deleted = true, "
-        + "r.changed = current_timestamp WHERE r.batchId = :batchId")
-    int markBatchAsDeleted(@Param("batchId") String batchId);
+    @Query("""
+        UPDATE RevocationBatchEntity r SET r.signedBatch = null, r.deleted = true,
+        r.changed = :currentTimestamp WHERE r.batchId = :batchId""")
+    int markBatchAsDeleted(@Param("batchId") String batchId,
+                           @Param("currentTimestamp") ZonedDateTime currentTimestamp);
 
     @Modifying
-    @Query("UPDATE RevocationBatchEntity r SET r.signedBatch = null, r.deleted = true, "
-        + "r.changed = current_timestamp WHERE r.deleted = false AND r.expires < :threshold")
-    int markExpiredBatchesAsDeleted(@Param("threshold") ZonedDateTime threshold);
+    @Query("""
+        UPDATE RevocationBatchEntity r SET r.signedBatch = null, r.deleted = true,
+        r.changed = :currentTimestamp WHERE r.deleted = false AND r.expires < :threshold""")
+    int markExpiredBatchesAsDeleted(@Param("threshold") ZonedDateTime threshold,
+                                    @Param("currentTimestamp") ZonedDateTime currentTimestamp);
 
     @Modifying
     @Query("DELETE FROM RevocationBatchEntity r WHERE r.deleted = true AND r.changed < :threshold")
