@@ -21,9 +21,9 @@
 package eu.europa.ec.dgc.gateway.restapi.dto.revocation;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.Pattern;
 import java.time.ZonedDateTime;
 import java.util.List;
-import javax.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
