@@ -31,13 +31,13 @@ import eu.europa.ec.dgc.gateway.model.RatValueset;
 import eu.europa.ec.dgc.gateway.model.Valueset;
 import eu.europa.ec.dgc.gateway.utils.DgcMdc;
 import feign.FeignException;
+import jakarta.annotation.PostConstruct;
+import jakarta.validation.ConstraintViolationException;
 import java.time.LocalDate;
 import java.time.ZonedDateTime;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Optional;
-import javax.annotation.PostConstruct;
-import javax.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
