@@ -23,9 +23,9 @@ package eu.europa.ec.dgc.gateway.service;
 import eu.europa.ec.dgc.gateway.entity.ValuesetEntity;
 import eu.europa.ec.dgc.gateway.repository.ValuesetRepository;
 import eu.europa.ec.dgc.gateway.utils.DgcMdc;
+import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
-import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
