@@ -36,6 +36,7 @@ import eu.europa.ec.dgc.gateway.repository.ValuesetRepository;
 import feign.FeignException;
 import feign.Request;
 import feign.RequestTemplate;
+import jakarta.validation.ConstraintViolationException;
 import java.time.LocalDate;
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoUnit;
@@ -43,7 +44,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import javax.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
